a bit code optimizing on bloom
